Extract Options from bogo::main.
